Horton Edits
One of the things I did was scale down the size of the plots. The default plot size when knitting to pdf is too large (in my opinion). This doesn't seem to be a problem in html, but in pdf, the result is images that appear far from where they are referenced.